# api-swagger.yaml
openapi: 3.0.0
info:
  description: Edu attend API
  version: '1.0.0'
  title: Edu attend API
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: trhoangthien1999@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'http://localhost:4000'
...
paths:
...
  /students:
    get:
      tags:
        - students
      summary: Get all students
      operationId: getStudents
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Get all students success'
                  data:
                    type: object
                    properties:
                      students:
                        type: array
                        items:
                          $ref: '#/components/schemas/Student'
        '401':
          description: Unauthorized
    post:
      tags:
        - students
      summary: Create student
      operationId: createStudent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentBody'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '401':
          description: Unauthorized
  /students/{id}:
    get:
      tags:
        - students
      summary: Get student detail
      operationId: getStudentDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '401':
          description: Unauthorized
    patch:
      tags:
        - students
      summary: Update student
      operationId: updateStudent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentBody'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '401':
          description: Unauthorized
    delete:
      tags:
        - students
      summary: Delete student
      operationId: deleteStudent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Delete student success'
        '401':
          description: Unauthorized
...
  /subjects:
    get:
      tags:
        - subjects
      summary: Get all subjects
      operationId: getSubjects
      security:
        - BearerAuth: []
      parameters:
        - name: school_id
          in: query
          description: Filter by school ID
          required: false
          schema:
            type: string
            format: MongoId
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Items per page
          required: false
          schema:
            type: integer
            default: 10
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Get all subjects success'
                  data:
                    type: object
                    properties:
                      subjects:
                        type: array
                        items:
                          $ref: '#/components/schemas/Subject'
        '401':
          description: Unauthorized
    post:
      tags:
        - subjects
      summary: Create subject
      operationId: createSubject
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectBody'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '401':
          description: Unauthorized
  /subjects/{id}:
    get:
      tags:
        - subjects
      summary: Get subject detail
      operationId: getSubjectDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '401':
          description: Unauthorized
    patch:
      tags:
        - subjects
      summary: Update subject
      operationId: updateSubject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectBody'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '401':
          description: Unauthorized
    delete:
      tags:
        - subjects
      summary: Delete subject
      operationId: deleteSubject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Delete subject success'
        '401':
          description: Unauthorized
  /timetables:
    get:
      tags:
        - timetables
      summary: Get all timetables
      operationId: getTimetables
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Get all timetables success'
                  data:
                    type: object
                    properties:
                      timetables:
                        type: array
                        items:
                          $ref: '#/components/schemas/Timetable'
        '401':
          description: Unauthorized
    post:
      tags:
        - timetables
      summary: Create timetable
      operationId: createTimetable
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimetableBody'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Timetable'
        '401':
          description: Unauthorized
  /timetables/{id}:
    get:
      tags:
        - timetables
      summary: Get timetable detail
      operationId: getTimetableDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Timetable'
        '401':
          description: Unauthorized
    patch:
      tags:
        - timetables
      summary: Update timetable
      operationId: updateTimetable
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimetableBody'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Timetable'
        '401':
          description: Unauthorized
    delete:
      tags:
        - timetables
      summary: Delete timetable
      operationId: deleteTimetable
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Delete timetable success'
        '401':
          description: Unauthorized
...
components:
  schemas:
...
    StudentBody:
      type: object
      required:
        - school_id
        - student_code
        - full_name
        - class_id
        - academic_year_id
      properties:
        school_id:
          type: string
          format: MongoId
        student_code:
          type: string
        full_name:
          type: string
        class_id:
          type: string
          format: MongoId
        academic_year_id:
          type: string
          format: MongoId
        status:
          type: integer
          enum: [0, 1, 2, 3] # ENROLLED, GRADUATED, DROPPED_OUT, SUSPENDED
        date_of_birth:
          type: string
          format: ISO8601
        address:
          type: string
        gender:
          type: string
        phone:
          type: string
        scholarship_percentage:
          type: number
        scholarship_description:
          type: string
    Student:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
        student_code:
          type: string
        full_name:
          type: string
        class_id:
          type: string
          format: MongoId
        school_id:
          type: string
          format: MongoId
        status:
          type: integer
        created_at:
          type: string
          format: ISO8601
        updated_at:
          type: string
          format: ISO8601
...
    SubjectBody:
      type: object
      required:
        - name
        - code
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
        credits:
          type: number
    Subject:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
        name:
          type: string
        code:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: ISO8601
        updated_at:
          type: string
          format: ISO8601
    TimetableBody:
      type: object
      required:
        - class_id
        - academic_year_id
        - start_date
        - end_date
        - schedule
      properties:
        class_id:
          type: string
          format: MongoId
        academic_year_id:
          type: string
          format: MongoId
        term:
          type: string
        start_date:
          type: string
          format: ISO8601
        end_date:
          type: string
          format: ISO8601
        schedule:
          type: array
          items:
            type: object
            required:
              - day_of_week
              - period
              - subject_id
            properties:
              day_of_week:
                type: string
              period:
                type: number
              subject_id:
                type: string
                format: MongoId
              teacher_id:
                type: string
                format: MongoId
              room:
                type: string
    Timetable:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
        class_id:
          type: string
          format: MongoId
        academic_year_id:
          type: string
          format: MongoId
        start_date:
          type: string
          format: ISO8601
        end_date:
          type: string
          format: ISO8601
        schedule:
          type: array
          items:
            type: object
        created_at:
          type: string
          format: ISO8601
        updated_at:
          type: string
          format: ISO8601
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT